kalshi assistant v4/requirements.txt
# Core Dependencies
requests
pandas
numpy
cryptography
python-dotenv
websockets
aiohttp
certifi
rich

# GUI (High-Performance GPU Accelerated)
dearpygui
Pillow
pystray
